add nexus-networking crate
move `plumb_service_firewall_rules` and its dependencies
