+++ nexus/networking/src/lib.rs
@@ -1,0 +1,12 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! Networking functionality shared between Nexus proper and its background
+//! tasks or sagas.
+
+mod firewall_rules;
+mod sled_client;
+
+pub use firewall_rules::*;
+pub use sled_client::*;
